generate/requirement.txt
py-solc-x==2.0.4